+++ ecal/core/src/readwrite/ecal_writer.h
@@ -72,8 +72,8 @@ namespace eCAL
     bool ShmSetBufferCount(long buffering_);
     bool ShmEnableZeroCopy(bool state_);
 
-    bool ShmSetAcknowledgeTimeout(int acknowledge_timeout_ms_);
-    int  ShmGetAcknowledgeTimeout();
+    bool ShmSetAcknowledgeTimeout(long long acknowledge_timeout_ms_);
+    long long  ShmGetAcknowledgeTimeout();
 
     bool AddEventCallback(eCAL_Publisher_Event type_, PubEventCallbackT callback_);
     bool RemEventCallback(eCAL_Publisher_Event type_);
@@ -135,7 +135,7 @@ namespace eCAL
 
     size_t             m_buffering_shm;
     bool               m_zero_copy;
-    int                m_acknowledge_timeout_ms;
+    long long          m_acknowledge_timeout_ms;
 
     std::atomic<bool>  m_connected;
     typedef Util::CExpMap<std::string, bool> ConnectedMapT;
